move IbcActionWithHandler into component mod
