+++ crates/core/component/ibc/src/lib.rs
@@ -8,6 +8,8 @@
 #[cfg_attr(docsrs, doc(cfg(feature = "component")))]
 #[cfg(feature = "component")]
 pub mod component;
+#[cfg(feature = "component")]
+pub use component::ibc_action_with_handler::IbcActionWithHandler;
 
 pub mod genesis;
 mod ibc_action;
@@ -15,7 +17,7 @@ mod ibc_token;
 pub mod params;
 mod version;
 
-pub use ibc_action::{IbcAction, IbcActionWithHandler};
+pub use ibc_action::IbcAction;
 pub use ibc_token::IbcToken;
 
 #[cfg_attr(docsrs, doc(cfg(feature = "component")))]
